api/requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
Flask-Migrate
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
MarkupSafe==2.1.2
pipreqs==0.4.11
psycopg2==2.9.5
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
six==1.16.0
SQLAlchemy
SQLAlchemy-Utils
typing_extensions
urllib3==1.26.14
Werkzeug==2.2.3
yarg==0.1.9
zipp==3.14.0
gunicorn